Fixed data load on init issue.

+++ lib/models/tea.dart
@@ -1,6 +1,6 @@
 class Tea {
   final String name;
-  final int strength;
+  final double strength;
 
   Tea({required this.name, required this.strength});
 }
